Fix the realloc issue in coreboot
For realloc function, previously it will free the original buffer
firstly and then alloc a new buffer from heap, then it will copy
the content from the original buffer to the new buffer. This has
no problem if the new buffer region has no overlap with the
original buffer region. Once there is overlap between the original
buffer region and the new buffer region, the original buffer's
content may be modified by the alloc to store the buffer management
metadata before we copy the content to the new buffer. Then we will
get a new buffer with invalid data after realloc. Once this situation
occurs, we may meet a random boot failure such as a kernel panic etc.

Ori-Tracked-On: OAM-117293
Tracked-On: OAM-118547
Signed-off-by: jizhenlo <zhenlong.z.ji@intel.com>
Signed-off-by: Chen, Gang G <gang.g.chen@intel.com>